+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## []
+- Add paragraph copy button
+
 ## [0.10.1] - 2026-05-26
 - initial preview release